==> 2_Extract_10_K_data.ipynb <==
# 2. Extract the company names and details
-------------------
Group 3 , September 24, 2022
1. Gezhi Cheng, 
2. Haowei Lee, 
3. Ziyi Liu, 
4. VS Chaitanya Madduri

> <i>Description: This notebook focuses on analysing the 10-k/10-q files present in the gogole drive and tries to extract the list of companies along with their details </i>





<div class="alert alert-block alert-info">
    <b>Tip:</b> # Please run this notebook in the colab . 
</div> 

### Pre requisites: 
1. And add the shortcut of the drive link : https://drive.google.com/drive/folders/1tZP9A0hrAj8ptNP3VE9weYZ3WDn9jHic to your personal drive.

The above drive contains the 10-k fillings from the author 
Bill McDonald

Do to the huge files we have used our personal google drive folders to save the files.

Files:
Company_Details_analysis.csv - to store the 10-k files of the companies selected

### Output files:
1. And add the shortcut of the drive link :https://drive.google.com/drive/folders/1X4UdGsQiHVWSr63FRiz8rwOuWW5Ua8uI?usp=sharing to your personal drive.

Files:
Selected_10k_v2.csv - to store the 10-k files of the companies selected



## 1. Import Required Packages 


import pandas as pd
import numpy as np
import os
import zipfile


# Load the Drive helper and mount 
# Please allow the access to the google drive
from google.colab import drive
drive.mount('/content/drive')

## 2.Loading the files to the notebook

### 2.1 Extract the list of the zip files

# getting the list of the zip files in the main folder zip .  
files = [] # empty list to caputure the zip files
dir_path = "/content/drive/MyDrive/10-X_C_Zip" # Main path directory 
  # Iterate directory
for path in os.listdir(dir_path):
    # check if current path is a file
    if os.path.isfile(os.path.join(dir_path, path)):
        files.append(path) # appending the file names to the list

# Printing the number of the files
print(len(files)) 


# printing the zip file names 
files

# Based on our analysis we have decided to collect all the 10-k fillings between
# 2011-2021

files = [
  '10-X_C_2011-2015.zip',
 '10-X_C_2016-2020.zip',
 '10-X_C_2021.zip']

len(files)


DIR_PATH = "/content/drive/MyDrive/SPM_files/" # defining the path in the Google drive

# Profitability Analysis
company_list = pd.read_csv(DIR_PATH + "Company_Details_analysis.csv")

# if you are loading the file from local files
# company_list = pd.read_csv("Company_Details_analysis.csv")

company_list.head()

## 3. Creating nexted dictionaries for file location
- we wil use this dictionary to get location of all the 10-k filling we requre from the company list.
- Level 1 will store the main zip files eg : '10-X_C_2011-2015.zip'
- Level 2 wil store the company index numbers eg: 755001
- Level 3 will store the folder paths eg: '2011/QTR1/20110203_10-K_edgar_data_755001_0001193125-11-022690.txt'

final_dic = {}

# Extract the file names and fodler details for the selected companies
# please create the company folders in the destination folder.
drive_path = "/content/drive/MyDrive/10-X_C_Zip/"


for file_name in files:
  final_dic[file_name] = {} 
  for index in company_list['Index_Key'].unique():
    final_dic[file_name][index]= {}
  zip_filepath = drive_path + file_name
  zip = zipfile.ZipFile(drive_path + file_name)
  for index in company_list['Index_Key'].unique():
    final_dic[file_name][index]= [s for s in zip.namelist() if ('_' + str(index)+ '_' in s) & ('10-K_edgar' in s)]


final_dic.keys()

# Dictionary contains teh location of all the files that need to be extrated
final_dic['10-X_C_2011-2015.zip'].keys()

final_dic.keys() # printing main dictionary keys

# Printing an example company key
final_dic['10-X_C_2016-2020.zip'][884614]

## 3. Extract the text data , company name  and details form the zip files

3.1 Extract text content to Dataframe

# A master dictionary to save the company details
# which details of the company are specified in th elist
columns = ['Company_Key', 'Text_data', 'Quarter_details',"Year"] 
# creating the empty dataframe with columns
master_dataframe = pd.DataFrame(columns= columns) 


# please create the company folders in the destination folder.
drive_path = "/content/drive/MyDrive/10-X_C_Zip/" # the input folder

#looping around the zip files 
for file_name in files: 
  # zip file handler  
  zip = zipfile.ZipFile(drive_path + file_name)
  # looping around the company index codes that we selected
  for company in company_list['Index_Key'].unique():
    # Checking if the path exists
    if final_dic[file_name][company]:
      # looping around all the file paths 
      for text_file_name in final_dic[file_name][company]: 
         # getting latest index of file to create the next index for entry
        index_val = master_dataframe.shape[0]
        try:
          # capturing the 10-k filling data and filepath
          master_dataframe.loc[index_val,"Text_data"] =  zip.read(text_file_name)
          master_dataframe.loc[index_val,"filename"] =  text_file_name
        except:
          # except block in case if there are errors in the code
          print("Name of the file :: {}".format(text_file_name))
    

final_dic.keys()

## 4.Extracting company details 

def extract_compnay_name (text_input):
   '''
   This function will extarct the name of the company from text
   input: 
    text_input: string which contains all the 10-k info
   return:
    name of the company
   '''
   file_data = str(text_input).split('\\n')
   for line in file_data:
     stripped_line = line.strip()
     if "COMPANY CONFORMED NAME:" in stripped_line :                
       splitted_str = stripped_line.split(":")  # split COMPANY CONFORMED NAME and real name
       company_name = splitted_str[1] 
       return company_name.strip() 

def extract_company_index (text_input):
   '''
    This function will extarct the name of the company from text
   input: 
    text_input: string which contains all the 10-k info
   return:
    company Index
   '''   
   file_data = str(text_input).split('\\n')
   for line in file_data:
     stripped_line = line.strip()
     if "CENTRAL INDEX KEY:" in stripped_line:
            
      splitted_str = stripped_line.split(":")  # split COMPANY CONFORMED NAME and real name
      company_index = splitted_str[1]
      return company_index.strip() 
      

### 4.1 Extract the Company name.

# Extract the company name
master_dataframe['company_name'] = master_dataframe['Text_data'].apply(lambda x: extract_compnay_name(x)) 

# code to remove special charaters from the name 
master_dataframe['company_name'] = master_dataframe['company_name'].str[6:-2]

### 4.2 Extract the 10-k filling year and quarter

# Extract the Year details from the file path 
master_dataframe['Year'] = master_dataframe['filename'].str.split("/").str[0]

# Extract the Quarter details from the file path 
master_dataframe['Quarter_details'] =  master_dataframe['filename'].str[5:9]

### 4.2 Extract the Company key

# Extract the company index code.
master_dataframe['Company_Key'] = master_dataframe['Text_data'].apply(lambda x: extract_company_index(x)) 

# Removing the special charaters from the code
master_dataframe['Company_Key'] = master_dataframe['Company_Key'].str[6:-2]

# type casting the key to integer
master_dataframe['Company_Key'] = master_dataframe['Company_Key'].astype(int)

# showing the top 5 records
master_dataframe.head()


#### Note:
For most of the 10-k fillings has the company code in the filename but it is not the case for all the files. SO in the below step we am comparing the company code from the filename to inside the file.


<div class="alert alert-block alert-info">
    <b>FYI:</b> Since there are many file in the zip files it is time consuming and colab is crashing sometimes to process all the files. So we are extracting the the files based on the file name. 
</div> 

# extracting the company code from the file 
master_dataframe['cik_filename'] = master_dataframe['filename'].str.split('_').str[4]

# showing the top 5 records
master_dataframe['cik_filename'] = master_dataframe['cik_filename'].astype(int)

master_dataframe.head(2)

# comparing of the company key from the file is same as the company key from text 
master_dataframe[master_dataframe['Company_Key'] != master_dataframe['cik_filename']]

- 79 cases there is an mismatch between the company key from the file and company key from text 

master_dataframe.shape

### 4.3 Creating an flag column to highlight the mismatch cases 

#Creating an flag column to highlioght the mismatch cases 
master_dataframe['actual_list'] = np.where(master_dataframe['Company_Key'].isin(company_list['Index_Key'].unique()), 1,0)

master_dataframe['actual_list'].value_counts()

master_dataframe.info()

master_dataframe.drop(['actual_list'], axis=1, inplace=True)

## 5. Exporting the files to the drive

# exporting the fikes to the storage system
master_dataframe.to_csv("/content/drive/MyDrive/SPM_files/Selected_10k_v2.csv", index=None)

# code to understan the size of the file 
from sys import getsizeof
print("Size of the dataframe : {} bytes".format(getsizeof(master_dataframe)))

## End of the Notebook
