==> 3_data_cleaning_sample.ipynb <==
# 3. Text data Cleaning
-------------------
Group 3 , September 28, 2022
1. Gezhi Cheng, 
2. Haowei Lee, 
3. Ziyi Liu, 
4. VS Chaitanya Madduri

> <i>Description: In this notebook the text data will be preprocessed</i>


<div class="alert alert-block alert-info">
    <b>FYI:</b> # Please run this notebook in the colab . 
</div> 

### Pre requisites: 
1. And add the shortcut of the drive link :https://drive.google.com/drive/folders/1X4UdGsQiHVWSr63FRiz8rwOuWW5Ua8uI?usp=sharing to your personal drive.


- As the we used colab computation engine and also the files are very large in size. We used our personal google drive folders to save the large data files.


Files:
Selected_10k_v2.csv - to store the 10-k files of the companies selected

### Output files:

Files:
df_final.csv - To store the stemmed results and word frequencies. 



## 1. Import Required Packages 

# Connecting to the google drive
from google.colab import drive
drive.mount('/content/drive')
from IPython.display import clear_output

import pandas as pd
import numpy as np
import nltk
import re 

from nltk.corpus import stopwords                         # Removing all the stopwords
from nltk.stem.porter import PorterStemmer                # Reducing words to base form

nltk.download('stopwords')

## 2.Loading the files to the notebook

### 2.1 Load the dataframe with 10-k text

DIR_PATH = "/content/drive/MyDrive/SPM_files/"

# Data frame with all the 10-k filligs from 2011-2022 for the 
# selecetd 92 companies
df = pd.read_csv(DIR_PATH + "Selected_10k_v2.csv")


df.head()

df.shape

### 2.2 Extracting the 10-k fillings 

Some times the dataframe contains a mix of 10-K an 10-Q files and we are making sure to process only 10-K.

df['filing_type'] = np.where(df['Text_data'].str.contains("10-K_edgar_data"), "10-K", "10-Q")

df['filing_type'].value_counts()

df = df[df['filing_type'] == "10-K"]

# Reseting the index
df.reset_index(drop=True, inplace=True)

## 3 Stemming the Text data

stemming is the process of reducing inflected (or sometimes derived) words to their word stem, base or root form—generally a written word form.

# Printing the first text record until 200 character
df['Text_data'][0][1000:1200]

# dividing the dataset into two chunks as for 300 records it takes around two 
#hours for processing 
# Incase we are doing a batchwise execution
# df2 = df[601:901].copy()
# df3 = df[901:].copy()


def data_clean(input_text):
    '''
    The function will do the following:
    1. Extract only the Aplhas numerical strings 
    2. Apply porter stemming which will convert a word to its root form . 
        for example : hidden will be converted to hide
    3. removes words less than 3 letters (making sure to avoid residual strings)
    Augments:
    input_text: the text to be stemmed
    
    return :
    test_process/input_text : cleaned text if there are any errors it will return the orginal text
    '''
    ps = PorterStemmer()
    try:
        test_process = re.sub('[^a-zA-Z0-9]', ' ', input_text)         # Removing special symbols like ... ! and keeping only text
        test_process = test_process.lower()                                     # Lower case
        test_process = test_process.split()                                     # string split into words
        test_process = [ps.stem(word) for word in test_process                  # reducing words to base form
              if (not word in set(stopwords.words('english')) ) and len(word)>3  ]
        test_process = " ".join(test_process)
        return test_process
    except:
        return input_text


### 3.1 Removing the header tags

df['Text_data'] = df['Text_data'].str.split("</Header>").str[1]

Note: As below step was time taking we have disturbuted the records and ran individually in our machines and joinined the results at the end.

 # Please note this particular step takes more than an hour long time to execute.
 # For 300 records it takes around 4h.  We have processes around 900 10-k fillings
df['Text_data_cleaned'] = df['Text_data'].apply(lambda x: data_clean(x) )

# It takes more memory to print whole text , so printing part of the text
df['Text_data_cleaned'][0][1000:1200]

# storing the resulted dataframe in a temporary file.
df[['Company_Key', 'Quarter_details', 'Year', 'filing_type',
       'Text_data_cleaned',  'compamy_name']].to_csv(DIR_PATH + "selected_10k_cleaned_overall.csv", index=None) 

       

## 4. Frequency Calculations

# incase if the execution has been stopped we will continue with the saved files

#  please use the 
DIR_PATH = "/content/drive/MyDrive/SPM_files/"

# sample data
df = pd.read_csv(DIR_PATH + "selected_10k_cleaned_overall.csv")


df.head()

# Key words for potential zombie compies (from case study)
operation_related = ["restructure", "reorganization", "reorganize", "reorganized", "reorganizing", "turnaround", "restructuring", "cost", "asset"]
credit_related = ["loan", "concession", "swap", "forgiveness", "moratorium", "bond", "covenant"]

def get_negative(data):
    '''
    The function calculates the frequency count of the negative words used on 
    the text 
    Augments:
    data: Stemmed text

    
    return :
    negative_percentage : negative words frequencey count
    '''

    DIR_PATH = "/content/drive/MyDrive/Strategy_Final/"
    negative_words = []
    with open(DIR_PATH + 'negative-words.txt', 'r') as f:
        for line in f:
            negative_words.append(line[:-1])
    # negative_words


    negative_count = 0

    words = data.split()

    for _ in words:
        if _ in negative_words:
            negative_count += 1

    negative_percentage = negative_count / len(words)

    return negative_percentage


def get_operation(data):
  '''
    The function calculates the frequency count of the operational
     words used on  the text 
    Augments:
    data: Stemmed text

    
    return :
    operation_count : operations words frequency count
    '''
    operation_count = 0

    words = data.split()

    for _ in words:
        if _ in operation_related:
            operation_count += 1
    
    return operation_count / len(words)

def get_credit(data):
  '''
    The function calculates the frequency count of the credit
     words used on  the text 
    Augments:
    data: Stemmed text

    
    return :
    operation_count : credit words frequency count
    '''
    credit_count = 0

    words = data.split()

    for _ in words:
        if _ in credit_related:
            credit_count += 1

    return credit_count / len(words)

df_final = df.copy()
#loading the saved files from the storage if the kernal breaks


df_final.head()

### 4.1 Calculation of the frequency count

# Calculating operating percentage
df_final["operation_percentage"] = df_final["Text_data_cleaned"].apply(get_operation)


# Calculating negative words percentage
df_final["negative_percentage"] = df_final["Text_data_cleaned"].apply(get_negative)

# Calculating credit words percentage
df_final["credit_percentage"] = df_final["Text_data_cleaned"].apply(get_credit)

df_final.head(2)

### 4.2 Calculation of the frequency count

df_final.describe()

df_final.head()

### 4.3 Filtering the records for company with most restructuring effects

Exploratory data analysis

df_final[(df_final['operation_percentage']>0.02) & (df_final['negative_percentage']>0.025) & (df_final['operation_percentage']>0.01) ]

fg = df_final[(df_final['operation_percentage']>0.0155) & (df_final['negative_percentage']>0.019) & (df_final['operation_percentage']>0.002769) ]

fg = df_final[(df_final['operation_percentage']>0.0155) & (df_final['negative_percentage']>0.019) & (df_final['operation_percentage']>0.002769) ]

# exporting the file for future use
DIR_PATH = "/content/drive/MyDrive/SPM_files/"
df_final.to_csv(DIR_PATH + "final_df.csv", index=None)

## End of the Notebook
