==> 4_calculate_profitability.ipynb <==
# 4. Calculate Profitability
-------------------
Group 3 , September 24, 2022
1. Gezhi Cheng, 
2. Haowei Lee, 
3. Ziyi Liu, 
4. VS Chaitanya Madduri

> <i>Description: The program in this notebook
- conducts data cleaning on wrds data
- merges company informations with processed wrds data
- calculates profitability and percentage change of profitability on consecutive three years
</i>


<div class="alert alert-block alert-info">
    <b>Tip:</b> # Please run this notebook in the colab . 
</div> 

### Pre requisites: 
1. And add the shortcut of the drive link :https://drive.google.com/drive/folders/1X4UdGsQiHVWSr63FRiz8rwOuWW5Ua8uI?usp=sharing to your personal drive.


Due to the huge files we have used our personal google drive folders to save the files.

Files:
- wrds_data.csv - Financial data downloaded from [WRDS](https://wrds-www.wharton.upenn.edu/login/?next=/pages/get-data/center-research-security-prices-crsp/annual-update/crspcompustat-merged/fundamentals-annual/)
- Company_Details_analysis.csv - Data storing company general informations

### Output files:

Files:
profitability_data.csv - Data storing profitability and percentage change in profitablilty


## Load dependencies

# Mount the program to the folder on Google Drive to get access to other files on cloud
from google.colab import drive
drive.mount('/content/drive')

# import python packages
import numpy as np
import pandas as pd
from typing import List
import warnings

## 1. Preprocess financial data from WRDS



### 1.1 Define constants

# Define financial terms as CONSTANTS
GROSS_PROFIT = "Gross Profit (Loss)"
REVENUE = "Revenue - Total"
GROSS_PROFIT_MARGIN = "Gross Profit Margin"
NET_INCOME = "Net Income (Loss)"
NET_PROFIT_MARGIN = "Net Profit Margin"
OPERATING_EXPENSES = "Operating Expenses Total"
OPERATING_PROFIT_MARGIN = "Operating Profit Margin"
ASSETS = "Assets - Total"
EQUITY = "Stockholders Equity - Total"
RETAINED_EARNINGS = "Retained Earnings"
ROA = "ROA"
ROE = "ROE"

# Define tha map for renaming a dataframe
RENAMING_MAP = {
                "CONM": "Company Name",
                "CIK": "CIK Number",
                "ACT": "Current Assets - Total",
                "AT": "Assets - Total",
                "CH": "Cash",
                "INVOFS": "Inventory/Stock - Other",
                "LCT": "Current Liabilities - Total",
                "LT": "Liabilities - Total",
                "SEQ": "Stockholders' Equity - Total",
                "COGS": "Cost of Goods Sold",
                "EPSPI": "Earnings Per Share (Basic) Including Extraordinary Items",
                "EPSPX": "Earnings Per Share (Basic) Excluding Extraordinary Items",
                "GP": "Gross Profit (Loss)",
                "NI": "Net Income (Loss)",
                "OPEPS": "Earnings Per Share from Operations",
                "OPITI": "Operating Income - Total",
                "REVT": "Revenue - Total",
                "SALE": "Sales/Turnover (Net)",
                "OANCF": "Operating Activities Net Cash Flow",
                "UNWCC": "Working Capital (Use) - Increase (Decrease) (Cash Flow)",
                "ACQINVT": "Acquired Assets - Inventory",
                "DT": "Total Debt Including Current",
                "FINCH": "Finance Division - Cash",
                "TEQ": "Stockholders Equity - Total",
                "XOPR": "Operating Expenses Total",
                "RE": "Retained Earnings"
                }

# make sure the key in the renamming map is in lower case
RENAMING_MAP_LOWER = {k.lower(): v for k, v in RENAMING_MAP.items()}


# define the path of the root folder and of the data
DIR_PATH = "/content/drive/MyDrive/SPM_files/"  # take the current folder as the root folder
WRDS_DATA_PATH = "wrds_data.csv"

LIMIT = 3  # define the numbers of rows for data exploration

### 1.2 Inspect WRDS data

# read in the WRDS data
df_wrds = pd.read_csv(DIR_PATH + WRDS_DATA_PATH)

# rename column names with the map defined in Constants part
df_wrds = df_wrds.rename(columns=RENAMING_MAP_LOWER)

# inspect the dimension and the first few records of the wrds data
print(df_wrds.shape)
df_wrds.head(LIMIT)

### 1.3 Calculate finantial indices

# Define functions to calculate financial indices
def get_gross_profit_margin(df):
    df[GROSS_PROFIT_MARGIN] = df[GROSS_PROFIT] / df[REVENUE]
    return df

def get_net_profit_margin(df):
    df[NET_PROFIT_MARGIN] = df[NET_INCOME] / df[REVENUE]
    return df

def get_operating_profit_margin(df):
    df[OPERATING_PROFIT_MARGIN] = (df[GROSS_PROFIT] - df[OPERATING_EXPENSES]) / df[REVENUE]
    return df

def get_roa(df):
    df[ROA] = df[NET_INCOME] / df[ASSETS]
    return df

def get_roe(df):
    df[ROE] = df[NET_INCOME] / (df[EQUITY] + df[RETAINED_EARNINGS])
    return df

# apply the functions define above to get finantial indices
df_wrds.pipe(get_gross_profit_margin)\
       .pipe(get_net_profit_margin)\
       .pipe(get_operating_profit_margin)\
       .pipe(get_roa)\
       .pipe(get_roe)

# inspect the data
df_wrds.head(LIMIT)

## 2. Preprocess the data of companyies' details 

# Define CONSTANTS
COMPANY_DATA_PATH = "Company_Details_analysis.csv"
TARGET_COLUMNS = ["Index_Key", "Class_number","Std_Indust_Class", 'Company_Name']

# read in data of companies details
df_company = pd.read_csv(DIR_PATH + COMPANY_DATA_PATH)

# keep only useful columns for further processing
df_company = df_company[TARGET_COLUMNS]
df_company.head(LIMIT)

##3. Merge data from WRDS and data of companies' details into one single dataframe

### 3.1 Merge and inspect the merged data

# to avoid irrelevant warning messages
warnings.filterwarnings("ignore")  

# define the column at where to merge two dataframes
MERGE_COLUMN = "CIK Number"

# rename the column of company index for the convenience when merging two data
df_company.rename(columns={"Index_Key": MERGE_COLUMN}, inplace=True)
df_merged = df_wrds.merge(df_company, on =MERGE_COLUMN)

# inspect the data and its dimensions
print(df_merged.shape)
df_merged.head(LIMIT)

## 3.2 Define functions and constants

# CONSTANTS
PROFITABILITY = "profitability"
PROFITABILITY_NEXT_1 = "profitability_next_1"
PROFITABILITY_NEXT_2 = "profitability_next_2"
PROFITABILITY_NEXT_3 = "profitability_next_3"
PERCENTAGE_CHANGE_PROFITABILITY_NEXT_1 = "change_next_1"
PERCENTAGE_CHANGE_PROFITABILITY_NEXT_2 = "change_next_2"
PERCENTAGE_CHANGE_PROFITABILITY_NEXT_3 = "change_next_3"

COMPANY = "Company_Key"
COMP_NEXT_1 = "company_next_1"
COMP_NEXT_2 = "company_next_2"
COMP_NEXT_3 = "company_next_3"


RENAMING_MAP_MERGED = {
    "CIK Number": "Company_Key",
    "Company Name": "company",
    "Std_Indust_Class": "industry",
    "fyear": "year",
    "Profitability": "profitability"
    }

financial_columns = [GROSS_PROFIT, REVENUE, GROSS_PROFIT_MARGIN, NET_INCOME, NET_PROFIT_MARGIN, OPERATING_EXPENSES, ASSETS, EQUITY, RETAINED_EARNINGS, ROA, ROE]

KEY_COLUMNS = ["Company_Key", "year", "profitability", "Company_Name"] + financial_columns

# define functions to calculate profitability and percentage changes
def get_profitability(df, ratio_columns:List = [OPERATING_PROFIT_MARGIN, NET_PROFIT_MARGIN, ROA, ROE, GROSS_PROFIT_MARGIN]):
    """
    This function calculates the proftability by taking average on given columns of financial ratios
    input: 
        df: pd.DataFrame
        ratio_columns: the specified columns of financial ratio
    return:
        original dataframe with a new profitability column
    """
    df_ratios = df.copy()[ratio_columns]
    df_ratios[PROFITABILITY] = df_ratios.sum(axis=1) / len(ratio_columns)
    df[PROFITABILITY] = df_ratios[PROFITABILITY]

    return df

def get_key_columns(df, renaming_map: dict, key_col: List):
    """
    This function renames columns of given dataframe and keep only relevant columns
    input: 
        df: pd.DataFrame
        renaming_map: the mapper for renaming
        key_col: the specified columns to keep
    return:
        dataframe with renamed and relevant columns
    """
    df = df.rename(columns=renaming_map)
    df_key_col = df.copy()[key_col]

    return df_key_col

def reorder_df(df, columns = ['Company_Key', 'year']):
    """
    This function reorders the records by given columns. By default, the function
    reorders by columns 'Company_Key' and 'year'
    input: 
        df: pd.DataFrame
        columns: the specified columns to reorder
    return:
        dataframe with reordered recoreds
    """
    
    return df.sort_values(columns)

def get_percentage_change(df, n_next_1:int = -1, n_next_2:int = -2, n_next_3:int = -3):
    """
    n: int - shift by number n of rows
    """

    df[PROFITABILITY_NEXT_1] = df[PROFITABILITY].shift(n_next_1)  # shift by n years
    df[PROFITABILITY_NEXT_2] = df[PROFITABILITY].shift(n_next_2)  # shift by n years
    df[PROFITABILITY_NEXT_3] = df[PROFITABILITY].shift(n_next_3)  # shift by n years

    df[COMP_NEXT_1] = df[COMPANY].shift(n_next_1)  # shift by n years
    df[COMP_NEXT_2] = df[COMPANY].shift(n_next_2)  # shift by n years
    df[COMP_NEXT_3] = df[COMPANY].shift(n_next_3)  # shift by n years

    # df = df.dropna()  # drop NaN after shifting

    df[PERCENTAGE_CHANGE_PROFITABILITY_NEXT_1] = (((df[PROFITABILITY_NEXT_1] - df[PROFITABILITY]) / df[PROFITABILITY] ) * 100).round(2)
    df[PERCENTAGE_CHANGE_PROFITABILITY_NEXT_2] = (((df[PROFITABILITY_NEXT_2] - df[PROFITABILITY]) / df[PROFITABILITY] ) * 100).round(2)
    df[PERCENTAGE_CHANGE_PROFITABILITY_NEXT_3] = (((df[PROFITABILITY_NEXT_3] - df[PROFITABILITY]) / df[PROFITABILITY] ) * 100).round(2)

    df[PROFITABILITY] = df[PROFITABILITY].round(4)
    df[PROFITABILITY_NEXT_1] = df[PROFITABILITY_NEXT_1].round(4)
    df[PROFITABILITY_NEXT_2] = df[PROFITABILITY_NEXT_2].round(4)
    df[PROFITABILITY_NEXT_3] = df[PROFITABILITY_NEXT_3].round(4)

    return df

#### 3.3 Inspect the result of profatibility

# apply the functions to calculate profitability
df_profitability = df_merged.pipe(get_profitability)\
                            .pipe(get_key_columns, RENAMING_MAP_MERGED, KEY_COLUMNS)\
                            .pipe(reorder_df)\
                            .pipe(get_percentage_change)

# specift the order of the columns
COLUMNS_ORDER = ["year", "Company_Key", 
                 "Company_Name",  "company_next_1","company_next_2", "company_next_3",
                 "profitability", "profitability_next_1", "profitability_next_2", "profitability_next_3",
                 "change_next_1", "change_next_2", "change_next_3"] + financial_columns

# reorder the dataframe        
df_profitability = df_profitability[COLUMNS_ORDER]
df_profitability.head(LIMIT)

def remove_missing_records(df):
    # define the filters to clear the records that mix up companies after shifing the columns
    filter_diff_company = (df[COMPANY] != df[COMP_NEXT_1]) | \
                          (df[COMPANY] != df[COMP_NEXT_2]) | \
                          (df[COMPANY] != df[COMP_NEXT_3])

    # df_profitability[filter_diff_company]
    df[PROFITABILITY][filter_diff_company] = np.NaN
    df.dropna(inplace=True)
    df = df.drop(columns=[COMP_NEXT_1, COMP_NEXT_2, COMP_NEXT_3])

    return df

# apply the function and inspect the result
df_profitability = df_profitability.pipe(remove_missing_records)

### Export data as a CSV file

EXPORTED_NAME = "profitability_data.csv"
df_profitability.to_csv(DIR_PATH + EXPORTED_NAME, index=False)

# test if the exported data work by reading in the data
# only uncomment if needed
# df_readin = pd.read_csv(DIR_PATH + EXPORTED_NAME)
# print({"rows": df_readin.shape[0], "columns": df_readin.shape[1]})
# df_readin

## End of this notebook
